Move and modularise guided tour state
