+++ client/layout/guided-tours/docs/examples/selectors/has-selected-site-premium-or-business-plan.js
@@ -11,7 +11,7 @@ import { getSitePlan } from 'state/sites/selectors';
 import { getSelectedSiteId } from 'state/ui/selectors';
 
 // NOTE: selector moved here because tour is no longer active and serves as example only
-// to use in a tour, move back to 'state/ui/guided-tours/contexts' (see commented out import above)
+// to use in a tour, move back to 'state/guided-tours/contexts' (see commented out import above)
 /**
  * Returns true if the selected site is on the Premium or Business plan
  *
