+++ client/layout/guided-tours/index.js
@@ -1,7 +1,6 @@
 /**
  * External dependencies
  */
-
 import React from 'react';
 import { connect } from 'react-redux';
 
@@ -9,7 +8,7 @@ import { connect } from 'react-redux';
  * Internal dependencies
  */
 import AsyncLoad from 'components/async-load';
-import { getGuidedTourState } from 'state/ui/guided-tours/selectors';
+import { getGuidedTourState } from 'state/guided-tours/selectors';
 
 function GuidedTours( { shouldShow } ) {
 	if ( ! shouldShow ) {
