+++ client/layout/guided-tours/tours/activity-log-tour/meta.js
@@ -7,7 +7,7 @@ import { isDesktop } from '@automattic/viewport';
  * Internal dependencies
  */
 import { and } from 'layout/guided-tours/utils';
-import { isNotNewUser } from 'state/ui/guided-tours/contexts';
+import { isNotNewUser } from 'state/guided-tours/contexts';
 
 export default {
 	name: 'activityLogTour',
