+++ client/layout/guided-tours/tours/gdocs-integration-tour/meta.js
@@ -2,7 +2,7 @@
  * Internal dependencies
  */
 import { and } from 'layout/guided-tours/utils';
-import { hasUserPastedFromGoogleDocs } from 'state/ui/guided-tours/contexts';
+import { hasUserPastedFromGoogleDocs } from 'state/guided-tours/contexts';
 import { isCurrentUserEmailVerified } from 'state/current-user/selectors';
 
 export default {
