+++ client/layout/guided-tours/tours/main-tour/meta.js
@@ -2,7 +2,7 @@
  * Internal dependencies
  */
 import { and } from 'layout/guided-tours/utils';
-import { isNewUser, isEnabled } from 'state/ui/guided-tours/contexts';
+import { isNewUser, isEnabled } from 'state/guided-tours/contexts';
 
 export default {
 	name: 'main',
