+++ client/state/guided-tours/actions.js
@@ -1,11 +1,11 @@
 /**
  * Internal dependencies
  */
-
 import { GUIDED_TOUR_UPDATE, GUIDED_TOUR_PAUSE, GUIDED_TOUR_RESUME } from 'state/action-types';
-
 import { savePreference } from 'state/preferences/actions';
 import { getPreference } from 'state/preferences/selectors';
+
+import 'state/guided-tours/init';
 
 export function quitGuidedTour( { tour, stepName, finished } ) {
 	const quitAction = {
