+++ client/state/guided-tours/reducer.js
@@ -1,15 +1,15 @@
 /**
  * External dependencies
  */
-
 import { omit } from 'lodash';
 
 /**
  * Internal dependencies
  */
 import { GUIDED_TOUR_UPDATE, GUIDED_TOUR_PAUSE, GUIDED_TOUR_RESUME } from 'state/action-types';
+import { withStorageKey } from 'state/utils';
 
-export function guidedTour( state = {}, action ) {
+export function guidedTours( state = {}, action ) {
 	switch ( action.type ) {
 		case GUIDED_TOUR_UPDATE:
 		case GUIDED_TOUR_PAUSE:
@@ -19,4 +19,4 @@ export function guidedTour( state = {}, action ) {
 	return state;
 }
 
-export default guidedTour;
+export default withStorageKey( 'guidedTours', guidedTours );
