+++ client/state/guided-tours/selectors/find-ongoing-tour.js
@@ -1,7 +1,6 @@
 /**
  * External dependencies
  */
-
 import { findLast } from 'lodash';
 
 /**
@@ -9,6 +8,8 @@ import { findLast } from 'lodash';
  */
 import { getActionLog } from 'state/ui/action-log/selectors';
 import { GUIDED_TOUR_UPDATE } from 'state/action-types';
+
+import 'state/guided-tours/init';
 
 /*
  * When applicable, returns the name of the tour that has been started and not
