+++ client/state/guided-tours/selectors/get-tours-history.js
@@ -3,4 +3,6 @@
  */
 import { getPreference } from 'state/preferences/selectors';
 
+import 'state/guided-tours/init';
+
 export default ( state ) => getPreference( state, 'guided-tours-history' );
