+++ client/state/guided-tours/test/fixtures/config.js
@@ -6,7 +6,7 @@ import { stubTrue } from 'lodash';
 /**
  * Internal dependencies
  */
-import { isNewUser } from 'state/ui/guided-tours/contexts';
+import { isNewUser } from 'state/guided-tours/contexts';
 
 export const MainTourMeta = {
 	name: 'main',
